Remove mara_defer
Code execution during VM stack unwinding is problematic as it can create
new stack frames. Stack frames are allocated linearly in an arena so
either new frames are overwritten or old frames, yet to be unwound, are
overwritten. Zones are also chained in the same manner.

Cleaning up of unmanaged resource should be done with tag on reference
instead.